+++ sandpack-client/src/clients/node/inject-scripts/consoleHook.ts
@@ -1,13 +1,14 @@
-/* eslint-disable @typescript-eslint/ban-ts-comment,no-console,@typescript-eslint/explicit-function-return-type, no-restricted-globals */
+/* eslint-disable @typescript-eslint/no-explicit-any,@typescript-eslint/ban-ts-comment,no-console,@typescript-eslint/explicit-function-return-type, no-restricted-globals */
 import Hook from "console-feed/lib/Hook";
-import { Decode } from "console-feed/lib/Transform";
+import { Encode } from "console-feed/lib/Transform";
 
 Hook(window.console, (log) => {
+  const encodeMessage = Encode(log) as any;
   parent.postMessage(
     {
       type: "console",
       codesandbox: true,
-      log: Decode(log),
+      log: Array.isArray(encodeMessage) ? encodeMessage[0] : encodeMessage,
     },
     "*"
   );