+++ sandpack-react/src/components/Preview/index.tsx
@@ -26,8 +26,6 @@ import {
 } from "../../styles/shared";
 import { SignOutIcon } from "../icons";
 
-import { PreviewProgress } from "./PreviewProgress";
-
 export interface PreviewProps {
   style?: React.CSSProperties;
   showNavigator?: boolean;
@@ -219,8 +217,6 @@ export const SandpackPreview = React.forwardRef<
             clientId={clientId}
             showOpenInCodeSandbox={showOpenInCodeSandbox}
           />
-
-          <PreviewProgress clientId={clientId} />
 
           {showSandpackErrorOverlay && <ErrorOverlay />}