+++ sandpack-react/src/styles/index.ts
@@ -12,11 +12,12 @@ import { createStitchesMock } from "./stitches-mock";
 export const THEME_PREFIX = "sp";
 
 const getNodeProcess = (): false | string | undefined => {
-  if (typeof process !== "undefined") {
-    return process.env.SANDPACK_BARE_COMPONENTS;
-  }
+  if (typeof process === "undefined") return false;
 
-  return false;
+  return (
+    process.env.SANDPACK_BARE_COMPONENTS ||
+    process.env.NEXT_PUBLIC_SANDPACK_BARE_COMPONENTS
+  );
 };
 
 /**