+++ sandpack-react/src/templates/node/vite-vue.ts
@@ -57,7 +57,7 @@ export default defineConfig({
     "/package.json": {
       code: JSON.stringify({
         scripts: {
-          dev: "vite",
+          dev: "vite --force",
           build: "vite build",
           preview: "vite preview",
         },