+++ sandpack-themes/rollup.config.js
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 const typescript = require("@rollup/plugin-typescript");
 
 const pkg = require("./package.json");