+++ website/codesandbox-theme-docs/CHANGELOG.md
@@ -3,6 +3,126 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.24](https://github.com/shuding/nextra/compare/v2.0.23...v2.0.24) (2023-03-02)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.23](https://github.com/shuding/nextra/compare/v2.0.22...v2.0.23) (2023-03-01)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.22](https://github.com/shuding/nextra/compare/v2.0.21...v2.0.22) (2023-03-01)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.21](https://github.com/shuding/nextra/compare/v2.0.20...v2.0.21) (2023-02-28)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.20](https://github.com/shuding/nextra/compare/v2.0.19...v2.0.20) (2023-02-28)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.19](https://github.com/shuding/nextra/compare/v2.0.18...v2.0.19) (2023-02-28)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.18](https://github.com/shuding/nextra/compare/v2.0.17...v2.0.18) (2023-02-28)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.17](https://github.com/shuding/nextra/compare/v2.0.16...v2.0.17) (2023-02-24)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.16](https://github.com/shuding/nextra/compare/v2.0.15...v2.0.16) (2023-02-23)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.15](https://github.com/shuding/nextra/compare/v2.0.14...v2.0.15) (2023-02-22)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.14](https://github.com/shuding/nextra/compare/v2.0.13...v2.0.14) (2023-02-22)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.13](https://github.com/shuding/nextra/compare/v2.0.12...v2.0.13) (2023-02-22)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.12](https://github.com/shuding/nextra/compare/v2.0.11...v2.0.12) (2023-02-22)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.11](https://github.com/shuding/nextra/compare/v2.0.10...v2.0.11) (2023-02-22)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
+## [2.0.10](https://github.com/shuding/nextra/compare/v2.0.9...v2.0.10) (2023-02-21)
+
+**Note:** Version bump only for package codesandbox-theme-docs
+
+
+
+
+
 ## [2.0.9](https://github.com/shuding/nextra/compare/v2.0.8...v2.0.9) (2023-02-20)
 
 **Note:** Version bump only for package codesandbox-theme-docs